[UPDATE] Better archi
